Refactor List.Item to ListItem
