+++ src/components/List.js
@@ -14,30 +14,10 @@ List.defaultProps = {
 
 List.displayName = 'List';
 
-export const ListItem = ({ tag: Tag, ...rest }) => <Tag {...rest} />;
-
-ListItem.propTypes = {
-  tag: CustomPropTypes.Tag,
-};
-
-ListItem.defaultProps = {
-  tag: 'li',
-};
-
-ListItem.displayName = 'List.Item';
-
 const StyleList = styled(List)`
   list-style-type: none;
   margin: 0;
   padding: 0;
 `;
-
-const StyleListItem = styled(ListItem)`
-  &:not(:last-child) {
-    margin-bottom: 0.5rem;
-  }
-`;
-
-StyleList.Item = StyleListItem;
 
 export default StyleList;
