+++ src/components/PhotoGrid.js
@@ -2,13 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import styled from 'styled-components';
 import List from './List';
+import ListItem from './ListItem';
 import GatsbyImage from './GatsbyImage';
 import * as CustomPropTypes from '../propTypes';
 import { pxToRem } from '../styles/utils';
 
 const breakpoint = pxToRem(800);
 
-const PrimaryPhotoGrid = styled(List.Item)`
+const PrimaryPhotoGrid = styled(ListItem)`
   @media (min-width: ${breakpoint}) {
     grid-column: span 2;
     grid-row: span 2;
@@ -23,9 +24,9 @@ const PhotoGrid = props => {
         <GatsbyImage sizes={primaryImage.sizes} />
       </PrimaryPhotoGrid>
       {additionalImages.map(image => (
-        <List.Item key={image.id}>
+        <ListItem key={image.id}>
           <GatsbyImage sizes={image.sizes} />
-        </List.Item>
+        </ListItem>
       ))}
     </List>
   );
@@ -47,11 +48,11 @@ export default styled(PhotoGrid)`
     grid-template-rows: repeat(2, ${pxToRem(200)});
   }
 
-  ${List.Item} {
+  ${ListItem} {
     margin: 0;
   }
 
-  ${List.Item} .gatsby-image-outer-wrapper,
+  ${ListItem} .gatsby-image-outer-wrapper,
   ${GatsbyImage} {
     height: 100%;
   }
